+++ NEWS.md
@@ -1,9 +1,18 @@
+### 0.1.3
+
+* Fixed the description file: NeedsCompilation is now set to "no".
+
+* Updated the reference to the main article.
+
+* Fixed the ptsr.link function. Now both derivatives corresponding to the identity link return a vector of ones with the same size of the function's argument.
+ 
 ### 0.1.2 
 
 * Fixed an error: Inside the function predict.ptsr, the argument xreg was not being passed to the internal function ".ptsr.predict.". 
 
 * Fixed an error: in the internal function ".ptsr.predict.", the MA lags were incorrectly specified. This was causing an error in the prediction for some models.
 
+* Release date: 	2022-02-08
 
 ### 0.1.1 
 
@@ -11,7 +20,7 @@
 
 * Added an internal function to calculate the numerical hessian. This function is only used if the inversion of the output of numDeriv::hessian fails.
 
-* Release date: 	2022-01-13
+* Release date: 	2022-02-01
 
 ### 0.1.0